# More efficient decoding of case classes for Scala 3